requirements.txt
ipykernel
pandas
nltk
scikit-learn
pyyaml
spacy
flask
requests
pandarallel
pyarrow
fastparquet